+++ src/app/art/art.page.ts
@@ -2,32 +2,35 @@ import { Component, ViewChild, effect } from '@angular/core';
 import { IonContent, IonicModule } from '@ionic/angular';
 import { Art } from '../models';
 import { DbService } from '../db.service';
-import { Router, RouterLink } from '@angular/router';
+import { Router, RouterLink, Scroll } from '@angular/router';
 import { CommonModule } from '@angular/common';
 import { ArtComponent } from './art.component';
 import { UiService } from '../ui.service';
 import { SearchComponent } from '../search/search.component';
 import { Network } from '@capacitor/network';
-import { NgxVirtualScrollModule } from '@lithiumjs/ngx-virtual-scroll';
+import { SkeletonArtComponent } from '../skeleton-art/skeleton-art.component';
+import { isWhiteSpace } from '../utils';
+import { CdkVirtualScrollViewport, ScrollingModule } from '@angular/cdk/scrolling';
 
 @Component({
   selector: 'app-arts',
   templateUrl: 'art.page.html',
   styleUrls: ['art.page.scss'],
   standalone: true,
-  imports: [IonicModule, RouterLink, CommonModule, NgxVirtualScrollModule,
-    ArtComponent, SearchComponent],
+  imports: [IonicModule, RouterLink, CommonModule, ScrollingModule,
+    ArtComponent, SearchComponent, SkeletonArtComponent],
 })
 export class ArtPage {
   showImage = false;
+  public busy = true;
+  public noArtMessage = 'No art was found';
   arts: Art[] = [];
-
-  @ViewChild(IonContent) ionContent!: IonContent;
+  minBufferPx = 1900;
+  @ViewChild(CdkVirtualScrollViewport) virtualScroll!: CdkVirtualScrollViewport;
 
   constructor(public db: DbService, private ui: UiService, private router: Router) {
     effect(() => {
-      console.log('scroll up')
-      this.ui.scrollUpContent('art', this.ionContent);
+      this.ui.scrollUp('art', this.virtualScroll);
     });
   }
 
@@ -36,11 +39,22 @@ export class ArtPage {
   }
 
   async ionViewDidEnter() {
-    const status = await Network.getStatus();
-    this.showImage = (status.connectionType == 'wifi');
-    if (this.arts.length == 0) {
-      this.update(undefined);
+    if (this.arts.length > 0) {
+      this.hack();
+      return;
     }
+
+    this.busy = true;
+    setTimeout(async () => {
+      try {
+        const status = await Network.getStatus();
+        this.showImage = (status.connectionType == 'wifi');
+        await this.update(undefined);
+      } finally {
+        this.busy = false;
+      }
+    }, 500);
+
   }
 
   handleInput(event: any) {
@@ -49,19 +63,26 @@ export class ArtPage {
 
   search(val: string | undefined | null) {
     if (!val) return;
-    this.ionContent.scrollToTop(100);
+    this.virtualScroll.scrollToOffset(0, 'smooth');
+    console.log(`Search for art "${val}"`);
+    this.noArtMessage = isWhiteSpace(val) ? `No art were found.` : `No art were found matching "${val}"`;
     this.update(val.toLowerCase());
+  }
+
+  private hack() {
+    // Hack to ensure tab view is updated on switch of tabs or when day is changed
+    this.minBufferPx = (this.minBufferPx == 1901) ? 1900 : 1901;
   }
 
   artTrackBy(index: number, art: Art) {
     return art.uid;
   }
 
-  async update(search: string | undefined) {
+  private async update(search: string | undefined) {
     this.arts = await this.db.findArts(search);
   }
 
   click(art: Art) {
-    this.router.navigate(['/art/'+ art.uid+'+Art']);
+    this.router.navigate(['/art/' + art.uid + '+Art']);
   }
 }