+++ changelog.md
@@ -1,4 +1,9 @@
 ## Changes
+
+### 1.4
+- Fix for region with non-US date formats
+- Fix for scrolling on Events and Camps page
+
 ### 1.3
 - Map of Restrooms added
 - fix for status bar color on ios in darkmode